requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
idna==3.10
PyMySQL==1.1.2
pyngrok==7.3.0
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
